adding some more item that will be selled, change ISBN data type to long on buku.java, fix some format to follow ISBN data type change,
to-do edit some more login, create register

+++ src/main/java/Classes/Buku.java
@@ -10,7 +10,7 @@ public class Buku extends Item {
     private String pengarang;
     private int tahunTerbit;
     private String penerbit;
-    private int ISBN;
+    private long ISBN;
     private int halaman;
     private String genre;
 
@@ -29,7 +29,7 @@ public class Buku extends Item {
      * @param genre         Genre buku
      */
     public Buku(String jenis, int stok, double harga, String judul, String pengarang, 
-                int tahunTerbit, String penerbit, int ISBN, int halaman, String genre) {
+                int tahunTerbit, String penerbit, long ISBN, int halaman, String genre) {
         super(jenis, stok, harga);
         this.judul = judul;
         this.pengarang = pengarang;
@@ -117,7 +117,7 @@ public class Buku extends Item {
      * 
      * @return ISBN buku.
      */
-    public int getISBN() {
+    public long getISBN() {
         return ISBN;
     }
